Make sure that lib can be started if host can't be resolved
If the StatsD collector is not up for some reason and the application can't resolve
the name by calling `:inet.gethostbyname(host)` then it prevents the entire application from starting.

This commit makes sure that the TelemetryMetricsStatsD can be started in the errored state
and depending on the `host_resolution_interval` it can be recovered or not.

If `host_resolution_interval` isn't passed, then TelemetryMetricsStatsD will not try to resolve the host again.
However, when it is passed, it can try to recover it in `host_resolution_interval` milliseconds.